+++ forui/CHANGELOG.md
@@ -1,7 +1,8 @@
 ## Next
 
 ### Changes
-* Change default `FTextFieldStyle`'s vertical content padding from `5` to `15`.
+* Change `FTextField` to be usable in `Form`s. 
+* Change `FTextFieldStyle`'s default vertical content padding from `5` to `15`.
 * Split exports in `forui.dart` into sub-libraries.
 
 ## Fixes